feat: UI tweaks to preprare for user testing
