+++ packages/theme-wizard-website/src/pages/components/[...slug].astro
@@ -29,6 +29,8 @@ const componentList = await Promise.all(Object.entries(components).map(async ([s
 }))
 
 const slug = Astro.params.slug as keyof typeof components
+const nlDesignSystemDocsUrl = new URL('https://nldesignsystem.nl')
+nlDesignSystemDocsUrl.pathname = slug
 const componentModulePromiseFn = components[slug]
 const componentModule = await componentModulePromiseFn()
 const { default: meta, ...storyExports } = componentModule
@@ -83,7 +85,12 @@ const storiesWithTokens = stories.map(([id, story]) => {
           <a href="/components">Componenten</a> /
           <span>{slug}</span>
         </nav>
+
         <h1>{slug}</h1>
+
+        <p>
+          Bekijk de component-documentatie <a href={nlDesignSystemDocsUrl.toString()} target="_blank">nldesignsystem.nl</a>.
+        </p>
 
         <div id="story-container" data-story-ids={JSON.stringify(storyIds)} data-component-id={slug}></div>
 
@@ -91,7 +98,7 @@ const storiesWithTokens = stories.map(([id, story]) => {
           <section>
             <h2>{story.name}</h2>
             <wizard-story-preview>
-              {/* theme-preview makes sure the CSS generated by theme-wizard config gets applied */}
+              {/* theme-preview makes sure the CSS generated by theme-wizard config gets applied after reset */}
               <wizard-reset-theme>
                 <wizard-preview-theme>
                   <div data-story-container={id}></div>
