+++ packages/theme-wizard-website/src/pages/index.astro
@@ -38,6 +38,10 @@ if (!currentTemplate) {
           <h2>Maak design keuzes</h2>
           <wizard-tokens-form></wizard-tokens-form>
         </section>
+        <section>
+          <h2>Download thema</h2>
+          <wizard-tokens-download></wizard-tokens-download>
+        </section>
       </div>
 
       <wizard-preview-picker slot="nav" templates={TEMPLATES_JSON} value={currentTemplate}></wizard-preview-picker>
